move gradient out to own file
